Add files via upload
https://cremerlab.github.io/hplc-py/index.html
hplc-py: A Python Utility For Rapid Quantification of Complex Chemical Chromatograms

Summary
High-Performance Liquid Chromatography (HPLC) and Gas Chromatography are analytical
techniques which allow for the quantitative characterization of the chemical components of
mixtures [Figure 1(A)]. Technological advancements in sample preparation and mechanical
automation have allowed HPLC to become a high-throughput tool (Broeckhoven et al., 2019;
Kaplitz et al., 2020) which poses new challenges for reproducible and rapid analysis of the
resulting chromatograms. Here we present hplc-py, a Python package that permits rapid and
reliable quantitation of component signals within a chromatogram for pipelined workflows.
This is achieved by a signal detection and quantitation algorithm which i) identifies windows
of time which contain peaks and ii) infers the parameters of a mixture of amplitude-weighted
skew-normal distributions which sum to reconstruct the observed signal. This approach is
particularly effective at deconvolving highly overlapping signals, allowing for precise absolute
quantitation of chemical constituents with similar chromatographic retention times.